+++ src/cli.js
@@ -8,6 +8,7 @@ var path = require( 'path' );
 var templateFsm = require( './template.fsm' );
 var releases = require( './gh' );
 var templates = require( './template' );
+var npm = require( './npm' );
 var create = templates.init( ROOT_PATH );
 
 function createVersion( owner, repo, version, item ) {
@@ -131,9 +132,13 @@ function installVersion( owner, repo, version, versions ) {
 		url = latest.url;
 	}
 	releases.download( ROOT_PATH, owner, repo, version, url )
-		.then( function( result ) {
-			return releases.extract( result );
-		} );
+		.then( onDownload );
+}
+
+function onDownload( result ) {
+	return releases.extract( result )
+		.then( npm.installModules )
+		.then( fs.movePackage );
 }
 
 function items( owner, repo, version ) {