+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.1.x
 
+### 0.1.2
+
+ * Added support for defining template npm dependencies
+ * Changed location of package.json files for templates to _support folder
+
 ### 0.1.1
 
  * Better error messages for missing template versions.